Rename and move items to break things up
