+++ src/napari_experimental/_widget.py
@@ -8,10 +8,10 @@ from napari.layers import Layer
 from napari.utils.events import Event
 from qtpy.QtWidgets import QPushButton, QVBoxLayout, QWidget
 
-from napari_experimental.names_only import (
-    GroupLayer,
-    QtLayerNamesTreeModel,
-    QtLayerNamesTreeView,
+from napari_experimental.group_layer import GroupLayer
+from napari_experimental.group_layer_qt import (
+    QtGroupLayerModel,
+    QtGroupLayerView,
 )
 
 if TYPE_CHECKING:
@@ -29,12 +29,12 @@ class GroupLayerWidget(QWidget):
 
         self.viewer = viewer
 
-        self.layer_names = GroupLayer(self.global_layers)
-        self.layer_names_model = QtLayerNamesTreeModel(
-            self.layer_names, parent=self
+        self.group_layers = GroupLayer(self.global_layers)
+        self.group_layers_model = QtGroupLayerModel(
+            self.group_layers, parent=self
         )
-        self.layer_names_view = QtLayerNamesTreeView(
-            self.layer_names, parent=self
+        self.group_layers_view = QtGroupLayerView(
+            self.group_layers, parent=self
         )
 
         self.global_layers.events.inserted.connect(self._new_layer)
@@ -49,25 +49,25 @@ class GroupLayerWidget(QWidget):
         self.setLayout(QVBoxLayout())
         self.layout().addWidget(self.enter_debugger)
         self.layout().addWidget(self.add_group_button)
-        self.layout().addWidget(self.layer_names_view)
+        self.layout().addWidget(self.group_layers_view)
 
     def _new_layer_group(self) -> None:
         """ """
         # Still causes bugs when moving groups
         # inside other groups, to investigate!
-        self.layer_names.add_new_item()
+        self.group_layers.add_new_item()
 
     def _new_layer(self, event: Event) -> None:
         """
         :param event: index, value (layer that was inserted)
         """
-        self.layer_names.add_new_item(layer_ptr=event.value)
+        self.group_layers.add_new_item(layer_ptr=event.value)
 
     def _removed_layer(self, event: Tuple[int, Layer]) -> None:
         """
         :param event: index, value (layer that was removed)
         """
-        self.layer_names.remove_layer_item(layer_ptr=event.value)
+        self.group_layers.remove_layer_item(layer_ptr=event.value)
 
     def _enter_debug(self) -> None:
         """Placeholder method that allows the developer to
